[ci] Fix error when no junit files are passed to report generator
This resulted in the style being None and despite the report being
empty as well, we tried to send it to the agent and Python can't
send None as an argument.

To fix this return "success" style and also check whether the
report has any content before calling the agent.